% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACFPlot.R
\name{ACFPlot}
\alias{ACFPlot}
\title{Plot for the autocorrelation function of a given Markov chain.}
\usage{
ACFPlot(chain, lag.max = 10, ...)
}
\arguments{
\item{chain}{A vector of values from the Markov chain.}

\item{lag.max}{Maximum number of lags to be displayed in the plot. Dafault is 10.}

\item{...}{Additional graphical parameters to be passed to ggplot.}
}
\value{
The function returns the acf plot for a specified Markov chain.
}
\description{
\code{} The ACFPlot function returns the plot for the autocorrelation function (acf) of a given Markov chain.
}
\examples{
# Generate two sets of values
chain = rnorm(100)

# Produce the plot
ACFPlot(chain, ggtitle('acf'))

}

