\name{ConsensusMCMC-package}
\alias{ConsensusMCMC}
\alias{ConsensusMCMC}
\docType{package}
\title{
ConsensusMCMC package
}
\description{
An package performing the ConsensusMCMC algorithm for different univariate and multivariate target distributions.}
\details{
\tabular{ll}{
Package: \tab ConsensusMCMC\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-12-01\cr
License: \tab GPL-2\cr
}

}
\author{
Virginia Aglietti, Tamar Loach, Emilia Pompe

Maintainer: Virginia Aglietti <virginia.aglietti@stats.ox.ac.uk>, Tamar Loach <tamar.loach@stats.ox.ac.uk>, Emilia Pompe <emilia.pompe@stats.ox.ac.uk>.
}
\references{
@article{scott2016bayes,
  title={Bayes and big data: The consensus Monte Carlo algorithm},
  author={Scott, Steven L and Blocker, Alexander W and Bonassi, Fernando V and Chipman, Hugh A     and George, Edward I and McCulloch, Robert E},
  journal={International Journal of Management Science and Engineering Management},
  volume={11},
  number={2},
  pages={78--88},
  year={2016},
  publisher={Taylor \& Francis}
}
}
\keyword{ package }

