% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaMH.R
\name{GammaMH}
\alias{GammaMH}
\title{MH-Algorithm for a Univariate Gamma target distribution.}
\usage{
GammaMH(data, n, sigma, k_prior, theta_prior, s, x_0)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
\code{} The GammaMH function returns the Markov chain generated by a random walk Metropolis-Hasting 
algorithm having a univariate gamma distribution as target distribution.
}
\examples{
add(1, 1)
add(10, 1)
}

