% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistPlot.R
\name{HistPlot}
\alias{HistPlot}
\title{Density estimates of the histograms for a list of Markov chains.}
\usage{
HistPlot(list_of_vectors, method = NULL, burn_in = 0.1, size_line = 1,
  ...)
}
\arguments{
\item{list_of_vectors}{A list of Markov chains to be plotted. The number of element in the list can be greater or equal than one.}

\item{method}{A vector giving the different algorithms used to produce the Markov chains. The dafault value is NULL.}

\item{burn_in}{A proportion of the generated sample that need to be discarded in the plot. Need to be given as a percentage in decimal number of the algorithm iterations. The default value is 0.1.}

\item{size_line}{The size of the line in the plot. The default value is 1.}

\item{...}{Additional graphical parameters to be passed to ggplot.}
}
\value{
The function returns the density estimates for the histograms of the specified Markov chains.
}
\description{
\code{} The HistPlot function returns the density estimates for the histograms of a given list of Markov chains.
}
\examples{
# Generate two sets of values
chain1 = rnorm(100)
chain2 = rnorm(100)

# Produce the density plots
HistPlot(list(chain1, chain2), 
         method = c('Single machine', 'Multiple machines'), 
         burn_in = 0.2, 
         size_line = 2)

}

