% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalMH.R
\name{NormalMH}
\alias{NormalMH}
\title{MH-Algorithm for a Univariate Normal target distribution.}
\usage{
NormalMH(data, n, sigma, mean_prior, sigma_prior, sigma_known, s, x_0)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
\code{} The NormalMH function returns the Markov chain generated by a random walk Metropolis-Hasting 
algorithm having a univariate normal distribution as target distribution.
}
\examples{
add(1, 1)
add(10, 1)
}

