% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQPlot}
\alias{QQPlot}
\title{QQPlot for two Markov chains.}
\usage{
QQPlot(chain1, chain2, line = TRUE, size_point = 2, size_line = 1, ...)
}
\arguments{
\item{chain1}{The first Markov-chain to plot.}

\item{chain2}{The second Markov-chain to plot.}

\item{line}{Logical parameter. If TRUE, the line chain1 = chain2 is added to the qqplot. Default value is TRUE.}

\item{size_point}{The desired size of the points in the plot. Default value is 2.}

\item{size_line}{The desired size of the line in the plot. Default value is 1.}

\item{...}{Additional graphical parameters to be passed to ggplot.}
}
\value{
The function returns the qqplot for two specified Markov chains.
}
\description{
\code{} The QQPlot function returns the qqplot for two given Markov chains.
}
\examples{
# Generate two sets of values
chain1=rnorm(100)
chain2=rnorm(100)

# Produce the qqplot
QQPlot(chain1=rnorm(100), chain2=rnorm(100))

}

