% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TracePlot.R
\name{TracePlot}
\alias{TracePlot}
\title{Trace plots for a list of Markov chains.}
\usage{
TracePlot(list_of_vectors, method = NULL, burn_in = 0.1, size_line = 1,
  ...)
}
\arguments{
\item{list_of_vectors}{A list of Markov chains to be plotted. The number of element in the list can be greater or equal than one.}

\item{method}{A vector giving the different algorithms used to produce the Markov chains. The dafault value is NULL.}

\item{burn_in}{A proportion of the generated sample that need to be discarded in the plot. Need to be given as a percentage in decimal number of the algorithm iterations. The default value is 0.1.}

\item{size_line}{The size of the line in the plot. The default value is 1.}

\item{...}{Additional graphical parameters to be passed to ggplot.}
}
\value{
The function returns the qqplot for two specified Markov chains.
}
\description{
\code{} The TracePlot function returns the trace plots for a given list of Markov chains.
}
\examples{
# Generate two sets of values
chain1 = rnorm(100)
chain2 = rnorm(100)

# Produce the trace plots
TracePlot(list(chain1, chain2), 
         method = c('Single machine', 'Multiple machines'), 
         burn_in = 0.2)

}

