% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_function.R
\name{weightsComputation}
\alias{weightsComputation}
\title{Weights computation for Consensus MCMC algorithm}
\usage{
weightsComputation(df, method)
}
\arguments{
\item{df}{A dataframe derived from running \code{\link{BetaMH}} or \code{\link{NormalMH}} or \code{\link{GammaMH}} on the relative data.}

\item{method}{The method that needs to be used to aggregate the markov chains obtained from different machines. Need to be used when running in parallel.}
}
\value{
The functions returns the aggregated Markov chain from the Consensus MCMC.
}
\description{
\code{} The function computes the weights used for the Consensus MCMC algorithm. 
The computed weights are used to aggregate the Markov Chains obtained parallelizing the data on different machines.
}
\examples{

df = data.frame(lapply(lambda, function(y) y))
parallel_markov_chain = weightsComputation(df, method = "sample variance")

}

