\name{ConsensusMCMC-package}
\alias{ConsensusMCMC}
\alias{ConsensusMCMC}
\docType{package}
\title{
ConsensusMCMC package
}
\description{
An package performing the ConsensusMCMC algorithm for different univariate and multivariate target distributions.}
\details{
\tabular{ll}{
Package: \tab ConsensusMCMC\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-12-01\cr
License: \tab GPL-2\cr
}

}
\author{
Virginia Aglietti, Tamar Loach, Emilia Pompe

Maintainer: Virginia Aglietti <virginia.aglietti@stats.ox.ac.uk>, Tamar Loach <tamar.loach@stats.ox.ac.uk>, Emilia Pompe <emilia.pompe@stats.ox.ac.uk>.
}
\references{
Scott et al., 2016
}
\keyword{ package }

