\name{OxWaSP-package}
\alias{OxWaSP-package}
\alias{OxWaSP}
\docType{package}
\title{
OxWaSP example package
}
\description{
An example package showing compiling and calling simple C code, including
calling where there are external library dependencies and exotic compilers.}
\details{
\tabular{ll}{
Package: \tab OxWaSP\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-02-21\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Louis J. M. Aslett

Maintainer: Louis J. M. Aslett <aslett@stats.ox.ac.uk>
}
\references{
See OxWaSP slides for further information.
}
\keyword{ package }
\examples{
sq(2)
gslrng(5)
ssq(1:1000,1001:2)
}
