% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaMH.R
\name{GammaMH}
\alias{GammaMH}
\title{MH-Algorithm for a Univariate Gamma target distribution.}
\usage{
GammaMH(data, n, sigma, k_prior, theta_prior, s, x_0)
}
\arguments{
\item{data}{Observations distributed as a Poisson distribution.}

\item{n}{The required number of iterations for the MH-algorithm.}

\item{sigma}{The value of the standard deviation for the Gaussian proposal distribution of the random-walk MH-algorithm.}

\item{k_prior}{The prior value of the shape parameter for the Gamma prior distribution of the unknown parameter.}

\item{theta_prior}{The prior value of the rate parameter for the Gamma prior distribution of the unknown parameter.}

\item{s}{The number of machines on which to run the MH-algorithm. The value ranges between 1 and 4. Must be set different to 1 if running in parallel.}

\item{x_0}{The initial value for the unknown parameter of the Beta target distribution.}
}
\value{
The function returns a vector with the values of the Markov chain generated by the MH-algorithm.
}
\description{
\code{} The GammaMH function returns the Markov chain generated by a random walk Metropolis-Hasting 
algorithm having a univariate gamma distribution as target distribution.
}
\examples{
# Generate data
observations = rpois(10000, 4)

# Set the parameters for the function
nr_servers = 1 
n_iter = 100000
burn_in = 0.1*n_iter
sigma = 0.01
k_prior = 1.0
theta_prior = 4.0
x_0 = 4

# Run the function for the chosen parameters
markov_chain = GammaMH(observations, 
                       n_iter = n_iter, 
                       sigma = sigma, 
                       k_prior = k_prior, 
                       theta_prior = theta_prior, 
                       s = 1, 
                       x_0 = x_0) 


}

