% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_function.R
\name{weightsComputation}
\alias{weightsComputation}
\title{Weights computation for Consensus MCMC algorithm}
\usage{
weightsComputation(df, method)
}
\arguments{
\item{df}{A dataframe derived from running \code{\link{BetaMH}}}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
\code{} The function computes the weights used for the Consensus MCMC algorithm. 
The computed weights are used to aggregate the Markov Chains obtained parallelizing the data on different machines.
}
\examples{
add(1, 1)
add(10, 1)
}

